fix(sync-github): normalise with: overrides and derive paths for deploy shorthand
- Move normalizeWorkflowWith + deriveDeployPaths to setup-workflows.ts so
  sync-github and setup share the same logic
- Fix JSON.parse of unquoted TS keys (docs: true) by quoting identifiers
  before parsing
- Apply normalizeWorkflowWith + deriveDeployPaths in buildBatch so the
  sync correctly translates docs: true → type: docs + paths: docs/**
